05-10

+++ database/migrations/2023_04_22_112333_create_carts_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             // 这里应该还有个记录商品属性和规格名称的数组，参考淘宝
             // 'specs' => [ '布料：黄色', '尺寸：L' ],
             // $table->json('specs')->comment('商品规格');
-            $table->integer('number')->default(1)->comment('数量');
+            $table->unsignedSmallInteger('number')->default(1)->comment('数量');
             $table->tinyInteger('is_checked')->default(0)->comment('是否选中：0 否 1 是');
             $table->timestamps();
